// src/app/api/model/aggregateCastMember.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CastRole } from './castRole';


export interface AggregateCastMember { 
    adult?: boolean;
    gender?: number;
    id?: number;
    known_for_department?: string;
    name?: string;
    original_name?: string;
    popularity?: number;
    profile_path?: string | null;
    roles?: Array<CastRole>;
    total_episode_count?: number;
    order?: number;
}