// src/app/api/model/aggregateCrewMember.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CrewJob } from './crewJob';


export interface AggregateCrewMember { 
    adult?: boolean;
    gender?: number;
    id?: number;
    known_for_department?: string;
    name?: string;
    original_name?: string;
    popularity?: number;
    profile_path?: string | null;
    jobs?: Array<CrewJob>;
    department?: string;
    total_episode_count?: number;
}